Removed position: relative on col classes
We had a problem with popins that where inside cols. Their widths where limitedby the size of the col whereas we wanted them to stretch to their content
The precise case was on Course Management page, "changer l'horaire d'un cours".
I had to upgrade sass to make it run on my node (v10.11.0)